Performance improvements when executing scripts with a lot of command invokations in them
For example scripts with tight loops and a lot of variable manipulation. The arguments and options syntax for built-in commands is now cached so that the CIF file doesn't need to be reloaded and parsed every time. Such scripts now run over twice as fast with much less stress on the file server. This is implemented by storing the RCommandOptionList and RCommandArgumentList from the CIF file in the CThreadCommandConstructor (via the new ROptArgCache class).